requirements.txt
pyserial
python-dotenv
